Refactor AppState management and modularize effects
Extract and modularize `AppState` initialization and effects into
separate files for improved organization and maintainability. Replace
inline effect definitions with the `allEffects` utility function to
streamline setup. Update imports to reflect the new `@core/models`
structure.
