+++ src/app/core/app-state.store.ts
@@ -1,66 +1,16 @@
-import chroma, {Color} from "chroma-js";
-import {Palette} from "@palettes/models/palette.model";
-import {PaletteStyle} from "@palettes/models/palette-style.model";
-import {ColorSpace} from "@common/models/color-space.model";
-import {generatePalette} from "@palettes/helper/palette.helper";
 import {signalStore, withState} from "@ngrx/signals";
 import {converterEvents} from "./converter/converter.events";
-import {Events, on, withEffects, withReducer} from "@ngrx/signals/events";
+import {on, withEffects, withReducer} from "@ngrx/signals/events";
 import {colorChangedReducer, correctLightnessReducer, displayColorSpaceReducer, newRandomColorReducer, useAsBackgroundReducer, useBezierReducer} from "./converter/converter.reducers";
-import {inject} from "@angular/core";
-import {LocalStorage} from "@common/services/local-storage.service";
-import {map} from "rxjs";
 import {persistenceEvents} from "./common/persistence.events";
 import {palettesEvents} from "./palettes/palettes.events";
-import {ColorTheme} from "@common/models/color-theme.model";
 import {loadAppStateReducer} from "./common/persistence.reducers";
 import {commonEvents} from "./common/common.events";
 import {colorThemeChangedReducer} from "./common/common.reducers";
-import {ColorThemeService} from "@common/services/color-theme.service";
-import {colorThemeChangeEffect} from "./common/common.effects";
-import {saveStateEffect} from "./common/persistence.effects";
-import {colorChangedEffect, useAsBackgroundChangedEffect} from "./converter/converter.effects";
-import {createShades, createTints} from "@common/helpers/tints-and-shades.helper";
 import {newRandomPaletteReducer, paletteChangedReducer, restorePaletteReducer, styleChangedReducer, updatePaletteColorReducer, useRandomChangedReducer} from "@core/palettes/palettes.reducers";
-import {Router} from "@angular/router";
-import {navigateToPaletteIdEffect} from "@core/palettes/palettes.effects";
+import {initialState} from "@core/models/app-state.model";
+import {allEffects} from "@core/all-effects";
 
-
-export type AppState = {
-  // Converter related
-  currentColor: Color;
-  useAsBackground: boolean;
-  correctLightness: boolean;
-  useBezier: boolean;
-  displayColorSpace: ColorSpace;
-  tintColors: Color[];
-  shadeColors: Color[];
-
-  // Palette related
-  paletteStyle: PaletteStyle;
-  useRandomStyle: boolean;
-  currentPalette: Palette;
-
-  // Common
-  colorTheme: ColorTheme;
-}
-
-const initialColor = chroma.random();
-const initialState: AppState = {
-  currentColor: initialColor,
-  useAsBackground: false,
-  correctLightness: true,
-  useBezier: true,
-  displayColorSpace: "hsl",
-  tintColors: createTints(initialColor, true, true),
-  shadeColors: createShades(initialColor, true, true),
-
-  paletteStyle: "muted-analog-split",
-  useRandomStyle: false,
-  currentPalette: generatePalette("muted-analog-split"),
-
-  colorTheme: "system"
-};
 
 export const AppStateStore = signalStore(
   {providedIn: "root"},
@@ -81,44 +31,7 @@ export const AppStateStore = signalStore(
     on(palettesEvents.useRandomChanged, useRandomChangedReducer),
     on(palettesEvents.styleChanged, styleChangedReducer)
   ),
-  withEffects(
-    (
-      store,
-      events = inject(Events),
-      localStorageService = inject(LocalStorage),
-      themeService = inject(ColorThemeService),
-      router = inject(Router)
-    ) => ({
-      setColorTheme$: colorThemeChangeEffect(events, themeService),
-
-      setBackgroundColor$: useAsBackgroundChangedEffect(
-        events,
-        themeService,
-        store
-      ),
-
-      navigateToPalette$: navigateToPaletteIdEffect(
-        events,
-        router,
-        store
-      ),
-
-      colorChanged$: colorChangedEffect(events, themeService, store),
-
-      anyPersistableEvents$: events
-        .on(
-          commonEvents.colorThemeChanged,
-          converterEvents.newRandomColor,
-          converterEvents.colorChanged,
-          palettesEvents.paletteChanged
-        )
-        .pipe(
-          map(() => persistenceEvents.saveAppState())
-        ),
-
-      persist$: saveStateEffect(events, localStorageService, store)
-    })
-  )
+  withEffects(allEffects)
 );
 
 export type AppStateStore = InstanceType<typeof AppStateStore>;
